mdp + CSV

+++ src/Controller/FilmController.php
@@ -11,7 +11,15 @@ use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
 use App\Service\SearchFilm;
-
+use Symfony\Component\Serializer\Serializer;
+use Symfony\Component\Serializer\Encoder\CsvEncoder;
+use Symfony\Component\Serializer\Normalizer\ObjectNormalizer;
+use Symfony\Component\String\Slugger\SluggerInterface;
+use App\Form\ImportCsvForm;
+use Symfony\Component\Form\Extension\Core\Type\SubmitType;
+use Symfony\Component\Validator\Constraints\File;
+use Symfony\Component\Form\Extension\Core\Type\FileType;
+use Symfony\Component\Form\Extension\Core\Type\PasswordType;
 
 class FilmController extends AbstractController
 {
@@ -23,6 +31,8 @@ class FilmController extends AbstractController
            "films" => $film
        ]);
     }
+
+
 
     public function ajoutFilm(Request $request, ManagerRegistry $doctrine, SearchFilm $chercher):Response
     {
@@ -44,16 +54,81 @@ class FilmController extends AbstractController
 
         return $this->renderForm('film/ajout.html.twig', ['form'=> $form]);
     }
+
+
     public function details(ManagerRegistry $doctrine, int $id):Response
     {
        $entitymanager = $doctrine->getManager();
        $film= $entitymanager->getRepository(Film::class)->find($id);
        return $this->render('film/details.html.twig', ["film" => $film]);
     }
-    public function delete(ManagerRegistry $doctrine, Film $film){
-        $em = $doctrine->getManager();
-        $em->remove($film);
-        $em->flush();
-        return $this->redirect("/index");
+
+
+    public function delete(ManagerRegistry $doctrine, Film $film, Request $request){
+        $form = $this->createFormBuilder( [])
+        ->add('mdp', PasswordType::class)
+        ->add("submit", SubmitType::class)
+        ->getForm();
+    
+        $form->handleRequest($request);
+
+        if($form->isSubmitted() && $form->isValid()){
+            $data =$form->getData();
+            if($data['mdp']=='1234'){
+                $em = $doctrine->getManager();
+                $em->remove($film);
+                $em->flush();
+                return $this->redirect("/index");
+            }
+        }
+        return $this->renderForm('film/delete.html.twig',['form'=> $form]);
+    }
+
+
+    public function importCsv(Request $request, ManagerRegistry $doctrine):Response{
+        $serializer = new Serializer([new ObjectNormalizer()], [new CsvEncoder()]);
+
+        $form = $this->createFormBuilder( []) 
+        ->add('importCsv', FileType::class, [
+            'label' => 'import (CSV file)',
+            // unmapped means that this field is not associated to any entity property
+            'mapped' => false,
+            // make it optional so you don't have to re-upload the CSV file
+            // every time you edit the Product details
+            'required' => true,
+            // unmapped fields can't define their validation using annotations
+            // in the associated entity, so you can use the PHP constraint classes
+            'constraints' => [
+                new File([
+                    'maxSize' => '1024k',
+                    'mimeTypes' => [
+                        'application/csv'
+                    ]
+                ])
+            ],
+        ])
+        ->add("submit", SubmitType::class)
+        ->getForm();
+        $form -> handleRequest($request);
+        if($form->isSubmitted() && $form->isValid()){
+            $csvFile = $form->get("importCsv")->getData();
+            // decoding CSV contents
+            $data = $serializer->decode(file_get_contents($csvFile->getPathname()), 'csv');
+            foreach($data as $movie){
+                $film = new Film();
+                $film->setNom($movie['nom']);
+                $film->setNbreVotants(1);
+                $film->setNote($movie['note']);
+                $film->setDescription($movie['description']);
+                $entitymanager = $doctrine->getManager();
+                $entitymanager->persist($film);
+                $entitymanager->flush(); 
+
+            }
+            return $this->redirect("/index");
+        }
+
+        return $this->renderForm('film/importCsv.html.twig',['form'=> $form]);
     }
 }
+
